Add toggle TV on LCD display for VU+Solo4K
Todo: When going to standby the VU drivers go back to LCD. When going out
of standby the VU drivers do not restore live TV. Even when (re)starting
enigma2 somehow the proc is not correctly checked by the drivers. In fact
when the proc has LiveTV enabled the drivers should arrange to go back to
LiveTV when comming back from standby (or restart enigma2).