ScanSetup: Return the TerrestrialDescription in list and only when something is connected on NIM
The getNetworksForNim returns list of Networks for the specified NIM, but for DVB-T
the description was returned as string causing the DVB-T network to append on know_networks
letter by letter on initialization of ScanSimple class.

Also for DVB-T we didn't check if something was connected on NIM, so know_networks
were filled with DVB-T network but NIM was excluded from scanning as we don't scan
nims if nothing is connected. Now we return DVB-T network only is something is connected.

Signed-off-by: Mike Looijmans <milo@openpli.org>